get remote branches when getting remotes
