+++ pkg/commands/branch.go
@@ -17,6 +17,7 @@ type Branch struct {
 	Recency   string
 	Pushables string
 	Pullables string
+	Hash      string
 	Selected  bool
 }
