+++ pkg/commands/git.go
@@ -1061,40 +1061,3 @@ func (c *GitCommand) BeginInteractiveRebaseForCommit(commits []*Commit, commitIn
 func (c *GitCommand) SetUpstreamBranch(upstream string) error {
 	return c.OSCommand.RunCommand(fmt.Sprintf("git branch -u %s", upstream))
 }
-
-func (c *GitCommand) GetRemotes() ([]*Remote, error) {
-	goGitRemotes, err := c.Repo.Remotes()
-	if err != nil {
-		return nil, err
-	}
-
-	return nil, nil
-
-	remotes := make([]*Remote, len(goGitRemotes))
-
-	// TODO: consider including the goGitRemote itself
-	for i, goGitRemote := range goGitRemotes {
-		goGitBranches, err := goGitRemote.List(&gogit.ListOptions{})
-		if err != nil {
-			c.Log.Warn(err)
-			continue
-		}
-
-		branches := []*Branch{}
-		for _, goGitBranch := range goGitBranches {
-			// for now we're only getting branch references, not tags/notes/etc
-			if goGitBranch.Name().IsBranch() {
-				branches = append(branches, &Branch{
-					Name: goGitBranch.Name().String(),
-				})
-			}
-		}
-
-		remotes[i] = &Remote{
-			Name:     goGitRemote.Config().Name,
-			Urls:     goGitRemote.Config().URLs,
-			Branches: branches,
-		}
-	}
-	return remotes, nil
-}
